backend/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
dotenv==0.9.9
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
